1. Створив файл зі змінними і все туди переніс 2. Добавив функцію isToken() на перевірку токена в localStorage
