+++ src/scripts/variables.js
@@ -1,0 +1,18 @@
+// header BUTTONS
+const btnLogin = document.querySelector(".btn__login");
+const btnCreate = document.querySelector(".btn__create");
+const btnOut = document.querySelector(".btn__out");
+const headerLogo = document.querySelector(".header_logo");
+const headerLogoExit = document.querySelector(".header_logo__exit");
+
+// form LOGIN
+const formLogin = document.getElementById('form-login');
+const modalLogin = document.getElementById("exampleModal");
+const mainBlock = document.querySelector(".container--main");
+
+
+
+// Export BUTTONS
+export {btnLogin, btnCreate, btnOut, headerLogo, headerLogoExit};
+// Export form LOGIN
+export {formLogin, modalLogin, mainBlock};
